% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_grafico_casos_covid.R
\name{Dygraph_Server}
\alias{Dygraph_Server}
\title{Servidor: Gráficos de casos de covid por provincia y departamento}
\usage{
Dygraph_Server(id, data_sisa, amba_reducido_names, base_raster, area, part)
}
\arguments{
\item{id}{Module name}

\item{data_sisa}{Dataframe con los casos diarios reportados de COVID-19.}

\item{amba_reducido_names}{String. Vector con los nombres de los partidos
que conforman el AMBA.}

\item{base_raster}{Dataframe que lista todos los rasters y desagrega en
sus columnas características de interes, como si son rasters de
AMBA o Buenos Aires, si el cambio porcentual es semanal o prepandemia
o el momento del día que representan.}

\item{area}{El raster puede corresponder a Buenos Aires provincia o a AMBA.}

\item{part}{Partido de la provincia de Buenos Aires, seleccionado en otro
módulo.}
}
\value{
La fecha seleccionada por el usuario en el grafico de COVID-19
}
\description{
Servidor: Gráficos de casos de covid por provincia y departamento
}
