% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_t1_fecha_momento.R
\name{FechaMomento_Server}
\alias{FechaMomento_Server}
\title{Servidor: Carga y configuración de raster COVID-19 según selección del usuario}
\usage{
FechaMomento_Server(id, pool, mapa_zoom)
}
\arguments{
\item{id}{Identificador del módulo Shiny.}

\item{pool}{Objeto de conexión a base de datos, usado para cargar los raster desde la base.}

\item{mapa_zoom}{Reactive con el valor de zoom actual del mapa Leaflet, usado para mostrar u ocultar controles.}
}
\value{
Una lista de objetos \code{reactive()} que incluyen:
\describe{
\item{area}{Área seleccionada (AMBA o provincia).}
\item{fecha}{Fecha seleccionada por el usuario.}
\item{porcentaje}{Tipo de cambio porcentual (p.ej. semanal, prepandemia).}
\item{momento}{Momento del día (mañana, tarde, etc.).}
\item{opacity}{Opacidad seleccionada para el mapa.}
\item{imagen}{Raster generado a partir de las selecciones del usuario.}
\item{boton}{Trigger del botón para actualizar el mapa.}
\item{basemap}{Capa base seleccionada.}
}
}
\description{
Servidor: Carga y configuración de raster COVID-19 según selección del usuario
}
