% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_t1_fecha_momento.R
\name{FechaMomento_UI}
\alias{FechaMomento_UI}
\title{Elementos de interfaz de usuario de la barra flotante del tab de Mapa BsAs}
\usage{
FechaMomento_UI(id, base_raster)
}
\arguments{
\item{id}{Nombre del módulo.}

\item{base_raster}{Dataframe que lista todos los rasters y desagrega en sus columnas características de interés, como si son rasters de AMBA o Buenos Aires, si el cambio porcentual es semanal o prepandemia, o el momento del día que representan.}
}
\value{
Elementos de interfaz de usuario para la barra flotante en el tab Mapa Buenos Aires.
}
\description{
Esta función crea los elementos de la interfaz de usuario para la barra flotante
en el tab de Mapa Buenos Aires, que permite al usuario seleccionar opciones como el
mapa base, área de interés, fecha, momento del día y opciones de transparencia.
}
