% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fecha_momento.R
\name{FechaMomentoUI}
\alias{FechaMomentoUI}
\title{Elementos de interfaz de usuario de la barra flotante del tab de Mapa BsAs}
\usage{
FechaMomentoUI(id, base_raster)
}
\arguments{
\item{id}{Module name}

\item{base_raster}{Dataframe que lista todos los rasters y desagrega en
sus columnas características de interes, como si son rasters de
AMBA o Buenos Aires, si el cambio porcentual es semanal o prepandemia
o el momento del día que representan.}
}
\value{
Barra flotante del tab Mapa Buenos Aires.
}
\description{
Elementos de interfaz de usuario de la barra flotante del tab de Mapa BsAs
}
