% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_t2_reporte.R
\name{ggplot_casos_covid_doble}
\alias{ggplot_casos_covid_doble}
\title{Gráfico individual y comparativo de casos COVID}
\usage{
ggplot_casos_covid_doble(partido, fecha, partido_comunas)
}
\arguments{
\item{partido}{Nombre del partido o comuna (ej. "La Matanza" o "Comuna 1")}

\item{fecha}{Fecha a marcar con una línea vertical (formato "YYYY-MM-DD")}

\item{partido_comunas}{Vector de nombres de comunas de CABA (ej. paste("Comuna", 1:15))}
}
\value{
Una lista con dos ggplots: \code{grafico_partido} y \code{grafico_comparativo}
}
\description{
Gráfico individual y comparativo de casos COVID
}
