% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_t2_histograma_raster.R
\name{HistogramaRaster_Server}
\alias{HistogramaRaster_Server}
\title{Servidor: Histograma de la movilidad ciudadana por departamento}
\usage{
HistogramaRaster_Server(
  id,
  pool,
  imagen,
  amba_reducido_names,
  bsas_comunas,
  fecha,
  tipo_de_raster,
  partido,
  momento_dia
)
}
\arguments{
\item{id}{Module name}

\item{amba_reducido_names}{String. Vector con los nombres de los partidos
que conforman el AMBA.}

\item{bsas_comunas}{Mapa vectorial de la provincia de Buenos Aires simplifica
do.}

\item{fecha}{Fecha seleccionada.}

\item{tipo_de_raster}{String. Si el raster corresponde a el cambio porcentual
prepandemia (pc) o semanal (7dpc).}

\item{momento_dia}{Entero. Hora del dia representada en el raster, puede ser
0am, 8 am o 4 pm.}

\item{base_raster}{Dataframe que lista todos los rasters y desagrega en
sus columnas características de interes, como si son rasters de
AMBA o Buenos Aires, si el cambio porcentual es semanal o prepandemia
o el momento del día que representan.}

\item{area}{El raster puede corresponder a Buenos Aires provincia o a AMBA.}

\item{part}{Partido de la provincia de Buenos Aires, seleccionado en otro
módulo.}
}
\value{
Histograma de la movilidad ciudadana por partido, respetando la
escala de colores del raster.
}
\description{
Servidor: Histograma de la movilidad ciudadana por departamento
}
