% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mapa_baires.R
\name{MapaBaires_Server}
\alias{MapaBaires_Server}
\title{Servidor: Mapa raster de la provincia de Buenos Aires}
\usage{
MapaBaires_Server(
  id,
  bsas,
  imagen,
  area,
  opacidad,
  basemap,
  amba_reducido_names
)
}
\arguments{
\item{id}{Nombre del módulo. Este parámetro es necesario para la correcta integración del módulo de servidor dentro de la aplicación Shiny.}

\item{bsas}{Dataset de clase \code{sf} con los partidos de Buenos Aires, utilizado para agregar las geometrías de los municipios y su visualización en el mapa.}

\item{imagen}{Imagen raster que representa la movilidad ciudadana. Esta imagen es cargada y visualizada sobre el mapa.}

\item{area}{Indica el área geográfica de interés: puede ser 'provincia' o 'amba'. Dependiendo de este parámetro, el mapa se ajusta en términos de visualización.}

\item{opacidad}{Valor de opacidad del raster, que controla la transparencia de la imagen raster sobre el mapa.}

\item{basemap}{Estilo de mapa base para la visualización, como "calles" o "relieve", determinado por el usuario.}

\item{amba_reducido_names}{Vector de nombres de los partidos que conforman el área del AMBA (Área Metropolitana de Buenos Aires), usado para filtrar los datos geoespaciales.}
}
\value{
Un objeto \code{leaflet} renderizado con un mapa interactivo, incluyendo la capacidad de hacer zoom y desplazarse por el área seleccionada.
También retorna el valor de zoom cuando cambia durante la navegación en el mapa.
}
\description{
Esta función crea la lógica del servidor para renderizar el mapa interactivo de la provincia de Buenos Aires.
El mapa incluye un raster que representa la movilidad ciudadana y permite la visualización interactiva
de los cambios en la provincia. También incluye controles de mapa como la escala, medidas, y leyendas.
}
