% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mapa_baires.R
\name{MapaBaires_UI}
\alias{MapaBaires_UI}
\title{IU: Mapa raster de la provincia de Buenos Aires}
\usage{
MapaBaires_UI(id)
}
\arguments{
\item{id}{Module name}
}
\value{
posición del mapa leaflet de Buenos Aires con proxy.
También retorna el valor de zoom a medida que cambia durante la navegación.
}
\description{
Este mapa se encuentra el tab 1 correspondiente al
Mapa provincia de Buenos Aires.
}
