% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_t3_mapa_covid_departamentos.R
\name{MapaCovidDepartamentos_Server}
\alias{MapaCovidDepartamentos_Server}
\title{Servidor: Mapa prov. de Buenos Aires por departamento}
\usage{
MapaCovidDepartamentos_Server(id, pool, amba_caba)
}
\arguments{
\item{id}{Module name}

\item{amba_caba}{String. Vector con los nombres de los partidos
que conforman el AMBA sin las comunas de Provincia de Buenos Aires.}
}
\value{
Mapas y gráficos del tab 3.
}
\description{
Este mapa se encuentra en el tab 3 de GeoCovid app.
}
