% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mapa_covid_elecciones_input.R
\name{MapaCovidElecciones_Server}
\alias{MapaCovidElecciones_Server}
\title{Servidor: Selector de fecha}
\usage{
MapaCovidElecciones_Server(id, data_sisa, bsas, part)
}
\arguments{
\item{id}{Module name}

\item{data_sisa}{Dataframe con los casos diarios reportados de COVID-19.}

\item{bsas}{Mapa vectoria de la provincia de Buenos Aires simplifica
do.}

\item{part}{Partido de la provincia de Buenos Aires, seleccionado en otro
módulo.}
}
\value{
Fecha seleccionada por el usuario
}
\description{
Servidor: Selector de fecha
}
