% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_t2_mapa_partido.R
\name{MapaPartido_Server}
\alias{MapaPartido_Server}
\title{Servidor: Mapa raster por partido}
\usage{
MapaPartido_Server(
  id,
  pool,
  amba_reducido_names,
  bsas_comunas,
  area,
  fecha,
  tipo_de_raster,
  opacidad,
  part,
  momento_dia
)
}
\arguments{
\item{id}{Module name}

\item{amba_reducido_names}{String. Vector con los nombres de los partidos
que conforman el AMBA.}

\item{area}{El raster puede corresponder a Buenos Aires provincia o a AMBA.}

\item{fecha}{Fecha seleccionada.}

\item{tipo_de_raster}{String. Si el raster corresponde a el cambio porcentual
prepandemia (pc) o semanal (7dpc).}

\item{opacidad}{Double. Valor de opacidad del raster.}

\item{part}{Partido de la provincia de Buenos Aires, seleccionado en otro
módulo.}

\item{momento_dia}{Entero. Hora del dia representada en el raster, puede ser
0am, 8 am o 4 pm.}

\item{base_raster}{Dataframe que lista todos los rasters y desagrega en
sus columnas características de interes, como si son rasters de
AMBA o Buenos Aires, si el cambio porcentual es semanal o prepandemia
o el momento del día que representan.}

\item{bsas}{Dataset de clase sf con los partidos de Buenos Aires.}
}
\value{
Mapa raster por partido creado con leaflet.
}
\description{
Este mapa se encuentra en el tab 2, "por partido" de GeoCovid app
}
