% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_raster_loader.R
\name{rasterLoader}
\alias{rasterLoader}
\title{Carga de raster desde la base de datos}
\usage{
rasterLoader(pool, raster_data, area)
}
\arguments{
\item{pool}{Conexión al pool de base de datos.}

\item{raster_data}{Un dataframe que contiene la información del raster, incluyendo el nombre del archivo (filename).}

\item{area}{Una cadena de texto que indica el área de interés. Puede ser "amba" o "baires".}
}
\value{
Un objeto \code{rast} del paquete \code{terra}, que representa el archivo raster cargado desde la base de datos.
}
\description{
Esta función se encarga de cargar un archivo raster desde la base de datos según el área especificada
(AMBA o Buenos Aires). Utiliza una consulta SQL para recuperar el raster como un archivo binario
y lo guarda temporalmente como un archivo \code{.tif}, que luego es cargado como un objeto raster
utilizando el paquete \code{terra}.
}
