% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_reporte.R
\name{ReporteServer}
\alias{ReporteServer}
\title{Servidor: reporte}
\usage{
ReporteServer(
  id,
  part,
  fecha,
  base_raster,
  bsas,
  area,
  tipo_de_raster,
  opacidad,
  mapa_partido_manana,
  mapa_partido_tarde,
  mapa_partido_noche
)
}
\arguments{
\item{id}{Module name}

\item{part}{Partido de la provincia de Buenos Aires, seleccionado en otro
módulo.}

\item{fecha}{Fecha seleccionada.}

\item{base_raster}{Dataframe que lista todos los rasters y desagrega en
sus columnas características de interes, como si son rasters de
AMBA o Buenos Aires, si el cambio porcentual es semanal o prepandemia
o el momento del día que representan.}

\item{bsas}{Dataset de clase sf con los partidos de Buenos Aires.}

\item{area}{El raster puede corresponder a Buenos Aires provincia o a AMBA.}

\item{tipo_de_raster}{String. Si el raster corresponde a el cambio porcentual
prepandemia (pc) o semanal (7dpc).}

\item{opacidad}{Double. Valor de opacidad del raster.a}

\item{imagen}{Imagen raster representando la movilidad ciudadana.}
}
\value{
Devuelve un reporte en base a las selecciones en el tab2.
}
\description{
Este módulo se encuentra en el tab 2: 'Por partido' de GeoCovid app
}
