% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_t2_selector_mapa.R
\name{selectormapaServer}
\alias{selectormapaServer}
\title{Title}
\usage{
selectormapaServer(id, act_mapas, fecha, amba_reducido_names)
}
\arguments{
\item{id}{}

\item{act_mapas}{}

\item{fecha}{}
}
\description{
Title
}
