% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mapa_baires.R
\name{MapaBaires_Server}
\alias{MapaBaires_Server}
\title{Servidor: Mapa raster de la provincia de Buenos Aires}
\usage{
MapaBaires_Server(
  id,
  bsas,
  imagen,
  area,
  opacidad,
  basemap,
  amba_reducido_names
)
}
\arguments{
\item{id}{Module name}

\item{bsas}{Dataset de clase sf con los partidos de Buenos Aires.}

\item{imagen}{Imagen raster representando la movilidad ciudadana.}

\item{area}{El raster puede corresponder a Buenos Aires provincia o a AMBA.}

\item{opacidad}{Double. Valor de opacidad del raster.}

\item{basemap}{Estilo de mapa Leaflet. Puede mostrar las calles o el relieve.}

\item{amba_reducido_names}{String. Vector con los nombres de los partidos
que conforman el AMBA.}
}
\value{
Mapa leaflet de Buenos Aires con proxy. También retorna el valor de
zoom a medida que cambia durante la navegación.
}
\description{
Este mapa se encuentra el tab 1 correspondiente al
Mapa provincia de Buenos Aires.
}
