% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_t3_mapa_covid_departamentos.R
\name{MapaCovidDepartamentos_Server}
\alias{MapaCovidDepartamentos_Server}
\title{Server: Módulo de mapa raster por partido}
\usage{
MapaCovidDepartamentos_Server(id, pool, amba_caba)
}
\arguments{
\item{id}{\code{character}. Identificador del módulo Shiny.}

\item{pool}{Conexión a base de datos.}

\item{amba_caba}{\code{character}.}
}
\value{
Una lista de reactivos con el raster seleccionado (\code{mapa_partido}) y el nivel de zoom (\code{zoom_mapa_partido}).
}
\description{
Lógica del servidor para el módulo que muestra un mapa raster por partido.
Este módulo corresponde a la pestaña 2, "Por partido", de la aplicación GeoCovid.
}
