% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mapa_covid_departamentos.R
\name{MapaCovidDepartamentos_Server}
\alias{MapaCovidDepartamentos_Server}
\title{Servidor: Mapa prov. de Buenos Aires por departamento}
\usage{
MapaCovidDepartamentos_Server(id, data_sisa, base_raster, bsas, amba_caba)
}
\arguments{
\item{id}{Module name}

\item{data_sisa}{Dataframe con los casos diarios reportados de COVID-19.}

\item{base_raster}{Dataframe que lista todos los rasters y desagrega en
sus columnas características de interes, como si son rasters de
AMBA o Buenos Aires, si el cambio porcentual es semanal o prepandemia
o el momento del día que representan.}

\item{bsas}{Dataset de clase sf con los partidos de Buenos Aires.}

\item{amba_caba}{String. Vector con los nombres de los partidos
que conforman el AMBA sin las comunas de Provincia de Buenos Aires.}
}
\value{
Mapas y gráficos del tab 3.
}
\description{
Este mapa se encuentra en el tab 3 de GeoCovid app.
}
