% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_partidos_input.R
\name{Partidos_Server}
\alias{Partidos_Server}
\title{Servidor: Seleccion de departamentos de provincia de Buenos Aires con IU
dinámica.}
\usage{
Partidos_Server(id, bsas, amba_reducido_names)
}
\arguments{
\item{id}{Module name}

\item{bsas}{Dataset de clase sf con los partidos de Buenos Aires.}

\item{amba_reducido_names}{String. Vector con los nombres de los partidos
que conforman el AMBA.}
}
\value{
El area y el partido seleccionado.
}
\description{
Servidor: Seleccion de departamentos de provincia de Buenos Aires con IU
dinámica.
}
